code: изменен тип переменной минимума на вещественный

+++ main.cpp
@@ -18,7 +18,7 @@ int main()
         max = b;
     }
 
-    int min;
+    double min;
     if (a <= b)
     {
         min = a;
